Reject Promise with ToriiRedirectError vs a string
This is considered best practise because trying to debug a literal is
difficult and lacks good stack tracing tools. It also make programmatic
conditionals because they have to have knowledge of the magic string /
literal value used. This becomes especially problematic if they are
rejected with a POJO or other non-Error object.

- http://eslint.org/docs/rules/prefer-promise-reject-errors
- http://bluebirdjs.com/docs/warning-explanations.html#warning-a-promise-was-rejected-with-a-non-error